refactor: refactored code into original files
